board-project/src/main/java/com/project/boardproject/service/ArticleCommentService.java
package com.project.boardproject.service;

import com.project.boardproject.dto.ArticleCommentDto;
import com.project.boardproject.repository.ArticleCommentRepository;
import com.project.boardproject.repository.ArticleRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@RequiredArgsConstructor
@Transactional
@Service
public class ArticleCommentService {

    private final ArticleCommentRepository articleCommentRepository;
    private final ArticleRepository articleRepository;


    @Transactional(readOnly = true)
    public List<ArticleCommentDto> searchArticleComment(long articleId) {
        return List.of();
    }
}
